uint32_t used for timestamp handling
This data type is more appropriate than the unsigned long because the timestamps are carried as
32-bytes integers over network.

+++ src/capture.c
@@ -111,7 +111,7 @@ void got_packet(u_char *args, const struct pcap_pkthdr *header, const u_char *pa
   // Packet arrival time (us)
   double arrival_time;
   // Timestamp
-  unsigned long int timestamp;
+  uint32_t timestamp;
   
   // Sizes
   int size_ethernet;
